+++ Tests/Microsoft.AppCenter.Test.Functional/Analytics/TrackEventTest.cs
@@ -11,13 +11,8 @@ namespace Microsoft.AppCenter.Test.Functional.Analytics
 {
     using Analytics = Microsoft.AppCenter.Analytics.Analytics;
 
-    // TODO: Move to AppCenterCrashesTest.cs after resolving problem with tests
-    using Crashes = Microsoft.AppCenter.Crashes.Crashes;
-
     public class TrackEventTest
     {
-        private static object locker = new object();
-
         private readonly string _appSecret = Guid.NewGuid().ToString();
 
         [Fact]
@@ -38,7 +33,7 @@ namespace Microsoft.AppCenter.Test.Functional.Analytics
             Analytics.TrackEvent("Hello World");
 
             // Wait for processing event.
-            httpNetworkAdapter.HttpResponseTask.Wait(1000);
+            await httpNetworkAdapter.HttpResponseTask;
 
             // Verify. The start session can be in same batch as the event HTTP request so look for it inside.
             Assert.Equal("POST", httpNetworkAdapter.Method);
@@ -78,7 +73,7 @@ namespace Microsoft.AppCenter.Test.Functional.Analytics
             Analytics.TrackEvent("Hello World", properties);
 
             // Wait for processing event.
-            httpNetworkAdapter.HttpResponseTask.Wait(1000);
+            await httpNetworkAdapter.HttpResponseTask;
 
             // Verify. The start session can be in same batch as the event HTTP request so look for it inside.
             Assert.Equal("POST", httpNetworkAdapter.Method);
@@ -95,100 +90,6 @@ namespace Microsoft.AppCenter.Test.Functional.Analytics
                 Assert.NotNull(typedProperties.SelectToken($"[?(@.name == 'Key{i}' && @.value == 'Value{i}')]"));
             }
             Assert.Equal(1, httpNetworkAdapter.CallCount);
-
-        }
-
-
-        // TODO: Move following methods to AppCenterCrashesTest.cs after resolving issues with tests
-        [Fact]
-        public async Task EnableDisableTest()
-        {
-
-            // Start App Center.
-            AppCenter.LogLevel = LogLevel.Verbose;
-            AppCenter.Start(_appSecret, typeof(Crashes));
-
-            // Disable Appcenter.
-            AppCenter.SetEnabledAsync(false).Wait(1000);
-
-            // Verify disabled.
-            var isEnabled = AppCenter.IsEnabledAsync().Wait(1000);
-            var isEnabledCrashes = Crashes.IsEnabledAsync().Wait(1000);
-            Assert.False(isEnabled);
-            Assert.False(isEnabledCrashes);
-
-            // Restart SDK.
-            AppCenter.UnsetInstance();
-            Crashes.UnsetInstance();
-            AppCenter.LogLevel = LogLevel.Verbose;
-            AppCenter.Start(_appSecret, typeof(Crashes));
-
-            // Verify disabled.
-            var isEnabled2 = AppCenter.IsEnabledAsync().Wait(1000);
-            var isEnabledCrashes2 = Crashes.IsEnabledAsync().Wait(1000);
-            Assert.False(isEnabled2);
-            Assert.False(isEnabledCrashes2);
-
-            // Enable SDK.
-            AppCenter.SetEnabledAsync(true).Wait(1000);
-
-            // Verify enabled.
-            var isEnabled3 = AppCenter.IsEnabledAsync().Wait(1000);
-            var isEnabledCrashes3 = Crashes.IsEnabledAsync().Wait(1000);
-            Assert.True(isEnabled3);
-            Assert.True(isEnabledCrashes3);
-
-            // Restart SDK.
-            AppCenter.UnsetInstance();
-            Crashes.UnsetInstance();
-            AppCenter.LogLevel = LogLevel.Verbose;
-            AppCenter.Start(_appSecret, typeof(Crashes));
-
-            // Verify enabled.
-            var isEnabled4 = AppCenter.IsEnabledAsync().Wait(1000);
-            var isEnabledCrashes4 = Crashes.IsEnabledAsync().Wait(1000);
-            Assert.True(isEnabled4);
-            Assert.True(isEnabledCrashes4);
-
-        }
-
-        [Fact]
-        public async Task TrackErrorTest()
-        {
-            // Set up HttpNetworkAdapter.
-            var httpNetworkAdapter = new HttpNetworkAdapter(expectedLogType: "managedError");
-            DependencyConfiguration.HttpNetworkAdapter = httpNetworkAdapter;
-
-            // Start App Center.
-            AppCenter.LogLevel = LogLevel.Verbose;
-            AppCenter.Start(_appSecret, typeof(Crashes));
-
-            Crashes.TrackError(new Exception("The answert is 42"));
-            httpNetworkAdapter.HttpResponseTask.Wait(1000);
-            var events = httpNetworkAdapter.JsonContent.SelectTokens($"$.logs[?(@.type == 'managedError')]").ToList();
-            Assert.Equal(1, events.Count());
-
-        }
-
-        [Fact]
-        public async Task SetUserIdTest()
-        {
-            // Set up HttpNetworkAdapter.
-            var httpNetworkAdapter = new HttpNetworkAdapter(expectedLogType: "managedError");
-            DependencyConfiguration.HttpNetworkAdapter = httpNetworkAdapter;
-
-            // Start App Center.
-            AppCenter.LogLevel = LogLevel.Verbose;
-            AppCenter.Start(_appSecret, typeof(Crashes));
-            var userId = "I-am-test-user-id";
-            AppCenter.SetUserId(userId);
-
-            Crashes.TrackError(new Exception("The answert is 42"));
-            httpNetworkAdapter.HttpResponseTask.Wait(1000);
-            Console.WriteLine(httpNetworkAdapter.JsonContent.ToString());
-            var events = httpNetworkAdapter.JsonContent?.SelectTokens($"$.logs[?(@.type == 'managedError')]").ToList();
-            Assert.Equal(1, events?.Count());
-            Assert.Contains(userId, events?[0]);
         }
     }
 }
